Move this class into its own file.
