+++ plugin.php
@@ -43,6 +43,7 @@ if ( ! defined( 'Asset_Loader\\LOADED' ) ) {
 require_once __DIR__ . '/inc/namespace.php';
 require_once __DIR__ . '/inc/taxonomy.php';
 require_once __DIR__ . '/inc/class-users-controller.php';
+require_once __DIR__ . '/inc/class-insert-post-handler.php';
 require_once __DIR__ . '/inc/template.php';
 
 if ( is_admin() ) {
